+++ lib/src/mock_ip.dart
@@ -3,7 +3,7 @@ import 'constants.dart' show random;
 /// Generate random IPv4 address.
 ///
 /// [format] argument accepts integers in range from 0 to 255, separated
-/// by dots(`.`), which represent a group or octet in IP.
+/// by dots(`.`), which represent a group or octet in IPv4.
 /// Group can also be represented with `*`, which generates any
 /// number for particular group.
 ///
@@ -16,59 +16,62 @@ import 'constants.dart' show random;
 ///   mockIPv4() == mockIPv4('*.*.*.*')
 /// ```
 String mockIPv4([String format = '*.*.*.*']) {
-  var ip_list = format.split('.');
+  var ip = format.split('.');
 
-  ip_list.length != 4 ? throw ArgumentError('Invalid IPv4 format - Must contain four entries') : null;
+  if(ip.length != 4)
+    throw ArgumentError('Invalid IPv4 format - Must contain 4 groups');
 
-  ip_list.forEach((s){
-    if(s != '*') {
-      var myint = int.tryParse(s) ?? -1;
-      if(myint > 255 || myint < 0) throw ArgumentError('Integers must be in range of 0 and 255');
+  var _ip = ip.map((s){
+    if(s == '*'){
+      return '${random.nextInt(255 + 1).toString()}.';
     }
-  });
 
-  var ip = ip_list.map((s) =>
-  s.contains('*') ? "${random.nextInt(255 + 1).toString()}." : "$s."
-  ).join();
+    if(int.tryParse(s) >= 0 && int.tryParse(s) <= 255){
+      return '$s.';
+    }else{
+      throw ArgumentError('Integers must be in range of 0 and 255');
+    }
+  }).toList().join();
 
-  return ip.substring(0, ip.length - 1); // remove last dot character.
+  return _ip.substring(0, _ip.length - 1);
 }
-
-
 
 /// Generate random IPv6 address.
 ///
-/// [format] argument accepts base 16 decimals in range from 0 to 65536(Here this is converted to int range),
-/// separated by colons(`:`), which represent a group or octet in IP.
+/// [format] argument accepts integers in range from 0 to 65536,
+/// separated by colons(`:`), which represent a group or hextet in IPv6.
 /// Group can also be represented with `*`, which generates any
-/// hexa decimal number of 16 bits for a particular group.
+/// hexadecimal number of 16 bits for a particular group.
 ///
 /// Returns IP as [String].
 ///
 /// Example usage:
 /// ```dart
-///   mockIPv6(*:e331:93bf:*:a7c9:a63:*:*) // returns similar to IPv4
-///   mockIPv6(e1b3:7bae:*:3474:*:c0cc:462:c4b9)   // returns similar to IPv4
+///   mockIPv6('*:e331:93bf:*:a7c9:a63:*:*')
+///   mockIPv6('e1b3:7bae:*:3474:*:c0cc:462:c4b9')
 ///   mockIPv6() == mockIPv6('*:*:*:*:*:*:*:*')
 /// ```
 
 String mockIPv6([String format = '*:*:*:*:*:*:*:*']) {
-  var ip_list = format.split(':');
 
-  ip_list.length != 8 ? throw ArgumentError('Invalid IPv6 format - Must contain 8 entries') : null;
+  var ip = format.split(':');
 
-  ip_list.forEach((s){
-    if(s != '*') {
-      var myint = int.tryParse(s, radix: 16) ?? -1;
-      if(myint > 65535 || myint < 0) throw ArgumentError('Hexa decimal must be in range of 0 and 65535');
+  if(ip.length != 8)
+    throw ArgumentError('Invalid IPv6 format - Must contain 8 groups');
+
+  var _ip = ip.map((s){
+    if(s == '*'){
+      return '${random.nextInt(65535 + 1).toRadixString(16).padLeft(4, '0')}:';
     }
-  });
 
-  var ip = ip_list.map((s) =>
-  s.contains('*') ? "${random.nextInt(65535 + 1).toRadixString(16)}:" : "$s:"
-  ).join();
+    if(int.tryParse(s, radix:16) >= 0 && int.tryParse(s, radix:16) <= 65536){
+      return '${s.padLeft(4, '0')}:';
+    }else{
+      throw ArgumentError('Integers must be in range of 0 and 65536');
+    }
+  }).toList().join();
 
-  return ip.substring(0, ip.length - 1);
+  return _ip.substring(0, _ip.length - 1);
 }