+++ lib/src/mock_name.dart
@@ -5,12 +5,12 @@ import 'constants.dart' show random, femaleNames, maleNames;
 /// Returns `String` representing a first name.
 ///
 /// [gender] parameter can be set to `male` to return only
-/// male names or `female` to return only female names. Empty
-/// [gender] returns any male or any female name.
-/// Default is empty string.
+/// male names or `female` to return only female names or `any`
+/// to return any male or any female name.
+/// Default is `any`.
 ///
 /// Throws [ArgumentError] if gender is set to anything other
-/// than _'male'_ or _'female'_.
+/// than _'male'_ or _'female'_ or _'any'_.
 ///
 /// For list of names, check [names](https://github.com/PinkFrojd/mock_data/blob/master/assets/names.md).
 ///
@@ -20,24 +20,19 @@ import 'constants.dart' show random, femaleNames, maleNames;
 ///   mockName('male')   // returns male name.
 ///   mockName('female') // returns female name.
 /// ```
-String mockName([String gender = '']){
-
-  // Passed gender should be 'male' or 'female'
-  // so it must be contained in 'fe(male)' string.
-  // fem or mal or all other such words contained in female.
-  if(gender != '') {
-    if (!(gender == 'male' || gender == 'female')) {
-      throw ArgumentError('Invalid gender value');
-    }
-  }
+String mockName([String gender = 'any']){
 
   switch(gender){
     case 'male':
-      return maleNames.elementAt(random.nextInt(maleNames.length - 1));
+      return maleNames.elementAt(random.nextInt(maleNames.length));
     case 'female':
-      return femaleNames.elementAt(random.nextInt(femaleNames.length - 1));
-    default:
-      return maleNames.union(femaleNames).elementAt(
+      return femaleNames.elementAt(random.nextInt(femaleNames.length));
+    case 'any':
+      var allNames = maleNames.union(femaleNames).toList();
+      allNames.shuffle(random);
+      return allNames.elementAt(
           random.nextInt(maleNames.length + femaleNames.length));
+    default:
+      throw ArgumentError('Invalid gender value');
   }
 }