+++ CHANGELOG.md
@@ -1,12 +1,21 @@
+## 1.2.1
+
+- UPDATED: Added support for IPv6 instead of IPv4 only.
+- UPDATED: Added new color spaces for `mockColor`
+- UPDATED: Added new male and female names
+- IMPROVED: Added more tests
+- CHANGED: `mockName` defaults to _'any'_ instead of empty string
+- FIX: `mockName` test for gender changed to more valid one
+
 ## 1.2.0
 
-- NEW: mockColor, mockUrl and mockDate functions
+- NEW: `mockColor`, `mockUrl` and `mockDate` functions
 
 ## 1.1.0
 
-- NEW: mockName function
+- NEW: `mockName` function
 - Improved documentation
-- FIX: Removed last dot that was generated using mockIPv4 function
+- FIX: Removed last dot that was generated using _mockIPv4_ function
 
 ## 1.0.1